+++ plugin/ip2city/manifest.toml
@@ -1,0 +1,14 @@
+name = "ip2city"
+kind = "UDF"
+description = "UDF Expression Release"
+version = "1"
+license = ""
+sysVars = [
+
+]
+validate = "Validate"
+onInit = "OnInit"
+onShutdown = "OnShutdown"
+export = [
+    {extPoint="GetUserDefinedFuncClass", impl="GetUdfIp2city"},
+]
